pyproject.toml
[project]
name = "mcp-task-list"
version = "0.1.0"
description = "task list app using mcp"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.11.3",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
]
